Initial search support plugin implementation.
This is essentially a trimmed-down version of org.eclipse.search with:
   * dependencies (such as ltk.refactoring) pruned
   * UI surface area reduced (less options in general, and a streamlined result view in particular)
   * compiler warnings addressed
   * legacy "classic" search support removed
   * UI nits addressed (e.g., view context menu selection awareness)

It's important to note that this is a first cut.  Still to do:
   * integration with the search box (and with that the temporary search menu can be removed)
   * further vetting of ported bits (e.g., unwanted extension point schema should be removed and unneeded property definitions should be deleted)

In the long run we'll also want to see how it makes sense to integrate this effort with our existing model search engine support in core.  This will become relevant when we want to support semantic search.  The extension architecture here *may* suffice but we'll need to iterate.  In the meantime, I've confirmed that this is more than adequate for our more immediate cross-file search needs.
Review URL: http://codereview.chromium.org//8347012

git-svn-id: http://dart.googlecode.com/svn/branches/bleeding_edge@560 260f80e4-7a28-3924-810f-c04153c831b5